+++ release_notes.md
@@ -8,3 +8,5 @@
 - added simple coloring support for terminal print() method
 - add ReassignGuard to ```typing_tools``` which is a descriptor preventing re-assignment of a previously assigned attribute. This is for example useful to prevent accidental re-assignment of observables.
 - fixed bug where variable wasn't declared global in ```terminal``` module
+- added ```terminal``` support for windows (for now this does not support redirected input)
+- added create_bg_task() to ```async_tools```